+++ 3-checks/5-profile.sh
@@ -15,6 +15,11 @@ function run_profilers {
 platform=`../0-tools/platform.sh`
 hostname=`hostname`
 
+if [ ! -d words/$platform ]; then
+	echo 'ERROR: Run the script ./1-prepare.sh first.'
+    exit 1
+fi
+
 if [ -e profiling/$platform ]; then
 	rm -rf profiling/$platform/*
 else
